+++ includes/givewp-addons/items-mailchimp-for-wp.php
@@ -70,7 +70,7 @@ function ddw_tbexgive_aoitems_mailchimp_for_wp( $admin_bar ) {
 				'href'   => 'https://us2.admin.mailchimp.com/audience/',
 				'meta'   => array(
 					'target' => ddw_tbex_meta_target(),
-					'title'  => esc_attr__( 'My Mailchimp Audience (mailchimp.com)', 'toolbar-extras-givewp' ),
+					'title'  => esc_attr__( 'My Mailchimp Audience', 'toolbar-extras-givewp' ) . ' (mailchimp.com)',
 				)
 			)
 		);
@@ -83,7 +83,7 @@ function ddw_tbexgive_aoitems_mailchimp_for_wp( $admin_bar ) {
 				'href'   => 'https://us2.admin.mailchimp.com/campaigns/',
 				'meta'   => array(
 					'target' => ddw_tbex_meta_target(),
-					'title'  => esc_attr__( 'My Mailchimp Campaigns (mailchimp.com)', 'toolbar-extras-givewp' ),
+					'title'  => esc_attr__( 'My Mailchimp Campaigns', 'toolbar-extras-givewp' ) . ' (mailchimp.com)',
 				)
 			)
 		);