+++ CHANGES.md
@@ -17,7 +17,7 @@
 
 * *Plugin launch. Everything's new!*
 * New: 7 Free GiveWP extensions supported/ integrated
-* New: 15 Premium GiveWP extensions supported/ integrated
+* New: 19 Premium GiveWP extensions supported/ integrated
 
 
 ### ⚡ 0.9.0 - 2019-08-06